refactor:TOP-94 ToHotChip. StringUtil 을 core 모듈로 이동
