+++ core/compose-ui/src/main/java/com/example/compose_ui/component/chip/ToHotChip.kt
@@ -1,4 +1,4 @@
-package tht.feature.tohot.component.chip
+package com.example.compose_ui.component.chip
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Row
@@ -9,19 +9,31 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.compose_ui.component.text.p.ThtP2
-import tht.feature.tohot.userData
 
+/**
+ * decode 를 해서 전달주는 곳이 있고, 아닌곳이 있음
+ * decode 를 해서 전달주는 곳
+ * - MyPageInfoRows
+ *
+ * 모르는곳
+ * - ToHotUserInfoFullCard
+ *
+ * TODO: 모두 decode 해서 전달주도록?
+ */
 @Composable
 fun ToHotEmojiChip(
     modifier: Modifier = Modifier,
     content: String,
     emojiCode: String
 ) {
-    val code = Integer.decode("0x$emojiCode")
-    val emoji = String(Character.toChars(code))
+    val emoji = runCatching {
+        val code = Integer.decode("0x$emojiCode")
+        String(Character.toChars(code))
+    }.getOrNull() ?: emojiCode
     ToHotChip(
         modifier = modifier,
         item = "$emoji $content"
@@ -45,7 +57,9 @@ fun ToHotChip(
                 .padding(vertical = 4.dp, horizontal = 8.dp),
             text = item,
             fontWeight = FontWeight.Medium,
-            color = Color(0xFFF9FAFA)
+            color = Color(0xFFF9FAFA),
+            overflow = TextOverflow.Ellipsis,
+            maxLines = 1
         )
     }
 }
@@ -54,8 +68,8 @@ fun ToHotChip(
 @Preview(showBackground = true, backgroundColor = 0xFFFFFFFF, name = "emojiChip")
 private fun ToHotEmojiChipPreview() {
     ToHotEmojiChip(
-        content = userData.interests.list.last().title,
-        emojiCode = userData.interests.list.last().emojiCode
+        content = "content",
+        emojiCode = "1F3E0"
     )
 }
