+++ feature/signin/src/main/java/tht/feature/signin/signup/preferredgender/PreferredGenderFragment.kt
@@ -9,10 +9,10 @@ import kotlinx.coroutines.launch
 import tht.core.ui.delegate.viewBinding
 import tht.core.ui.extension.repeatOnStarted
 import tht.core.ui.extension.showToast
+import tht.core.ui.util.StringUtil
 import tht.feature.signin.databinding.FragmentPreferredGenderBinding
 import tht.feature.signin.signup.SignupRootBaseFragment
 import tht.feature.signin.signup.SignupRootViewModel
-import tht.feature.signin.util.StringUtil
 
 @AndroidEntryPoint
 class PreferredGenderFragment : SignupRootBaseFragment<PreferredGenderViewModel, FragmentPreferredGenderBinding>() {
